Added a player selection screen and a few other things.
Added a Menu class that can be used for the functionality (i.e. it
doesn't do any graphics) of a menu.

Added a PlayerInput class that we can use instead of tying ourself
directly to Unity's Input class. This might make some things easier to
deal with along the way.

Some of this probably needs some cleaning up and most of it will
definitely change over time.